requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
flask-googlemaps==0.4.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.0.1
future==0.18.2
htmlmin==0.1.12
itsdangerous==1.1.0
joblib==1.0.1
jsmin==2.2.2
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.4
mkdocs==1.1.2
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.1
mkdocs-minify-plugin==0.3.0
nltk==3.5
pymdown-extensions==8.1.1
pymongo==3.11.3
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
regex==2020.11.13
tqdm==4.59.0
Werkzeug==1.0.1
